ButtonHelp: refactor into its own component
